Added the bushes
