+++ scss/_FooterSection.scss
@@ -1,0 +1,30 @@
+.footer-section {
+
+    bottom: 0;
+    position: absolute;
+    width: 100%;
+
+  .content-container {
+    background-color: #1b5e20;
+    padding: 0px 10px;
+    margin-top: -2px !important;
+
+    @media (min-width: 768px) {
+      padding: 0px 100px;
+    }
+
+    h1 {
+      @apply text-base md:text-3xl;
+      padding-bottom: 10px;
+    }
+
+    p {
+      @apply text-xs md:text-lg;
+      padding-bottom: 30px;
+    }
+  }
+
+  .footerbush-image-container span {
+    display: block !important;
+  }
+}
