+++ scss/index.scss
@@ -3,6 +3,7 @@
 
 // import your custom scss files below here
 @import './AboutSection';
+@import './FooterSection';
 
 body {
   overflow-x: hidden;
